refactor(utils): extract completion helpers into mtui/completion.py
Move complete_choices and complete_choices_filelist verbatim out of
mtui/utils.py into a new mtui/completion.py module. utils.py keeps a
re-export shim for callers so this commit is byte-identical at the
public-API surface.

Second commit of the Phase 5b / C8 split (PLAN.md item 11).
